Use different functions depending on message version
